TST: Reorganize the utils test module.
