+++ act/utils/__init__.py
@@ -22,6 +22,7 @@ __getattr__, __dir__, __all__ = lazy.attach(
             'height_adjusted_temperature',
             'convert_to_potential_temp',
             'arm_site_location_search',
+            'DatastreamParserARM',
         ],
         'datetime_utils': [
             'dates_between',
